Fixed favicon, blog link, absolute urls, typos, github link
Went through and changed each index.html to feature relative URLs for easier local testing. Github link changed from old github account. I put the resources, workshops, and whitepapercircle files in their own directory. Only edited responsive.css of the css files in order to make the Sponsors box on the homepage  fit the formatting of the rest of the page. Fixed typos on the "What do We Do" section of the home page. Changed favicon to have a transparent background. Blog link on the main menu now works.